wib: move into folder and fix middleware cookie
